use provided result of to get types from gql query
